// src/interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface Category {
	_id: string
	name: string
	creationDate: string
	__v: number
	creationUser?: string
	operationType?: string
}

export interface Token {
	expiresIn: number
	token: string
}

export interface User {
	_id: string
	name: string
	lastname: string
	dni: number
	enabled: boolean
	email: string
	verified: boolean
	creationDate: string
}

export interface Book {
	_id: string
	name: string
	editorial: string
	course: string
	teacher: string
	author: string
	creationUser: User
	creationDate: string
	operationType: string
}